:bug: fix: validation 코드 변경 및 reissue api를 service -> jwtTokenUtils로 변경
