+++ src/main/java/com/springles/jwt/JwtTokenUtils.java
@@ -1,7 +1,11 @@
 package com.springles.jwt;
 
 import com.springles.domain.entity.RefreshToken;
+import com.springles.exception.CustomException;
+import com.springles.exception.constants.ErrorCode;
 import com.springles.repository.BlackListTokenRedisRepository;
+import com.springles.repository.MemberJpaRepository;
+import com.springles.repository.RefreshTokenRedisRepository;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.security.Keys;
 import lombok.extern.slf4j.Slf4j;
@@ -11,6 +15,7 @@ import org.springframework.stereotype.Component;
 import java.security.Key;
 import java.time.Instant;
 import java.util.Date;
+import java.util.Optional;
 import java.util.UUID;
 
 @Slf4j
@@ -21,15 +26,22 @@ public class JwtTokenUtils {
     private final JwtParser jwtParser;
     private final BlackListTokenRedisRepository blackListTokenRedisRepository;
 
+    private final MemberJpaRepository memberJpaRepository;
+    private final RefreshTokenRedisRepository refreshTokenRedisRepository;
+
     public JwtTokenUtils(
             @Value("${jwt.secret}") String jwtSecret,
-            BlackListTokenRedisRepository blackListTokenRedisRepository
+            BlackListTokenRedisRepository blackListTokenRedisRepository,
+            MemberJpaRepository memberJpaRepository,
+            RefreshTokenRedisRepository refreshTokenRedisRepository
     ) {
         this.singleKey = Keys.hmacShaKeyFor(jwtSecret.getBytes());
         this.jwtParser = Jwts.parserBuilder()
                 .setSigningKey(this.singleKey)
                 .build();
         this.blackListTokenRedisRepository = blackListTokenRedisRepository;
+        this.memberJpaRepository = memberJpaRepository;
+        this.refreshTokenRedisRepository = refreshTokenRedisRepository;
     }
 
     // accessToken 발급
@@ -37,7 +49,7 @@ public class JwtTokenUtils {
         Claims jwtClaims = Jwts.claims()
                 .setSubject(memberName)
                 .setIssuedAt(Date.from(Instant.now()))
-                .setExpiration(Date.from(Instant.now().plusSeconds(60 * 3))); // 3분(test용)
+                .setExpiration(Date.from(Instant.now().plusSeconds(60 * 1))); // 3분(test용)
         return Jwts.builder()
                 .setClaims(jwtClaims)
                 .signWith(singleKey)
@@ -49,26 +61,66 @@ public class JwtTokenUtils {
         return RefreshToken.builder()
                 .memberName(memberName)
                 .refreshToken(String.valueOf(UUID.randomUUID()))
-                .expiration(60 * 60 * 24 * 14L)   // 2주
+//                .expiration(60 * 60 * 24 * 14L)   // 2주
+                .expiration(60 * 2L)   // 2분(test용)
                 .build();
     }
 
-    public boolean validate(String accessToken) {
+    // accessToken 재발급
+    public String reissue(String refreshTokenId) {
+
+        // refreshToken이 있는지 확인
+        Optional<RefreshToken> optionalRefreshToken = refreshTokenRedisRepository.findById(refreshTokenId);
+        if(optionalRefreshToken.isEmpty()) {
+            log.warn("refresh token이 존재하지 않음");
+            throw new CustomException(ErrorCode.NO_JWT_TOKEN);
+        }
+
+        // refreshToken이 유효한지 확인
+        if(!memberJpaRepository.existsByMemberName(optionalRefreshToken.get().getMemberName())) {
+            throw new CustomException(ErrorCode.INVALID_REFRESH_TOKEN);
+        }
+
+        // accessToken 재발급
+        return generatedToken(optionalRefreshToken.get().getMemberName());
+    }
+
+//    public boolean validate(String accessToken) {
+//        try {
+//            jwtParser.parseClaimsJws(accessToken);
+//            return true;
+//        } catch (ExpiredJwtException e) {
+//            log.warn("유효기간이 만료된 token");
+//            return false;
+//        } catch (MalformedJwtException e) {
+//            log.warn("유효하지 않은 jwt 서명");
+//            return false;
+//        } catch (UnsupportedJwtException e) {
+//            log.warn("지원되지 않는 JWT 토큰");
+//            return false;
+//        } catch (IllegalArgumentException e) {
+//            log.warn("잘못된 JWT 토큰");
+//            return false;
+//        }
+//    }
+
+    public int validate(String accessToken) {
         try {
-            jwtParser.parseClaimsJws(accessToken);
-            return true;
+            jwtParser.parseClaimsJws(accessToken).getBody();
+            log.info("유효한 토큰");
+            return 1;
         } catch (ExpiredJwtException e) {
             log.warn("유효기간이 만료된 token");
-            return false;
+            return 2;
         } catch (MalformedJwtException e) {
             log.warn("유효하지 않은 jwt 서명");
-            return false;
+            return 0;
         } catch (UnsupportedJwtException e) {
             log.warn("지원되지 않는 JWT 토큰");
-            return false;
+            return 0;
         } catch (IllegalArgumentException e) {
             log.warn("잘못된 JWT 토큰");
-            return false;
+            return 0;
         }
     }
