+++ src/main/java/com/springles/service/RefreshTokenService.java
@@ -1,10 +1,10 @@
-package com.springles.service;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface RefreshTokenService {
-
-    String reissue(String refreshTokenId);
-
-}
+//package com.springles.service;
+//
+//import org.springframework.stereotype.Service;
+//
+//@Service
+//public interface RefreshTokenService {
+//
+//    String reissue(String refreshTokenId);
+//
+//}
